Support untracked cache with '--untracked-files=all' if configured
Untracked cache was originally designed to only work with
'-untracked-files=normal', but this is a concern for UI tooling
that wants to see "all" on a frequent basis, and the conditions
that prevented applicability to the "all" mode no longer seem to
apply.

The disqualification of untracked cache is a particular problem on
Windows with the defaulted fscache, where the introduction of
fsmonitor can make the first and costly directory-iteration happen
in untracked file detection, single-threaded, rather than in
preload-index on multiple threads. Improving the performance of a
"normal" 'git status' run with fsmonitor can make
'git status --untracked-files=all' perform much worse.

Here we align the supported directory flags for the stored
untracked cache data with the git config. If a user specifies an
'--untracked-files=' commandline parameter that does not align
with their 'status.showuntrackedfiles' config value, then the
untracked cache will be ignored - as it is for other unsupported
situations like when a pathspec is specified.

If the previously stored flags no longer match the current
configuration, but the currently-applicable flags do match the
current configuration, then the previously stored untracked cache
data is discarded.

For most users there will be no change in behavior. Users who
need '--untracked-files=all' to perform well will have the option
of setting "status.showuntrackedfiles" to "all".

Users who need '--untracked-files=all' to perform well for their
tooling AND prefer to avoid the verbosity of "all" when running
git status explicitly... are out of luck for now.

Users who set "status.showuntrackedfiles" to "all" and yet
most frequently explicitly call 'git status --untracked-files=all'
(and use the untracked cache) are the only users who will be
explicitly disadvantaged by this change. These should be
vanishingly rare, if there are any.