tests: t3505 incorrectly spills cherry-picking state
In t3505, some tests that check for cherry-pick failures forget to clean up
the cherry-picking state. Subsequent tests completed "accidentally" because
"git checkout" removed the cherry-picking state as a side-effect. As this
behavior of "git checkout" will change to instead abort, the tests need to
change.

Ensure tests that check for cherry-pick failure don't accidentally leave an
in-progress cherry-picking state.