tests: t2400 redundantly checks switching to a branch that is rebasing
One test in t2400 checks that git checkout is allowed to check out a branch
that is currently being rebased in the same worktree. The intent was to
check "locking" of the branch for that worktree. As we will be disallowing
git checkout during a rebase, however, this check will no longer work.

Checking out the branch that you are rebasing, while you are rebasing it,
is not a very obviously useful operation: If you later complete the rebase
it will have become a no-op, and if you instead quit, then you will have
thrown away any remaining commits in the rebase sequence.

Remove this test.