tests: t2030 relies on checkout -f bulldozing merge state
In t2030 the rerere tests sometimes leave a "merging" state between tests,
and a subsequent test then starts with "git checkout -f" to remove both
changes in the index, and the "merging" state. We will be changing
checkout to refuse to run when there is a merging state and instead suggest
the user "--quit".

Update tests that rely on "git checkout -f" to remove merging state, so that
they instead explicitly remove it with "git merge --quit".